Add nix flake for local development and gemfile
Installing jekyll, webbrick, bundler, and other dependencies while
mucking around environment variables is an option for local web
development, but I personally found it unpleasant when I was working
with the Morea framework. (A friend of mine straight-up couldn't get it
to work.)

To attempt to rectify that issue for myself, I've added a nix flake
which will allow the creation of an ephemeral shell that will set the
necessary environment variables and install the dependencies. Should
someone else wish to get the dependencies installed for development,
they need only use the Nix Package Manager with experimental features.
(I'm really doing this for myself, but everyone is free to take
advantage of this as they see fit.)